fix: Invalid behaviour
